fix(protolens): stop seven silent-wrong-output paths
Tier A of docs/protolens/audit-quality.md. Each item produced a
plausible-looking wrong answer rather than an error, which is the worst
failure mode for a tool whose job is to say what some bytes mean.

- a text blob with one bad byte opened as an empty document, because
  the encoder has no return type and leaves its buffer untouched
- batch export to stdout never flushed, so a full disk exited 0 with a
  truncated stream
- an export refused by the document's own inference still exited 0 and
  wrote a file; only --load-overrides refusals were reported
- a MessageSet extension id narrowed with `as u32` could name a valid
  but different type
- an unreadable descriptor set hashed as empty, so :save wrote a digest
  matching nothing and :restore warned about a phantom mismatch
- a disconnected input channel was reported as a clean exit

Field numbers turned out to be bounded already: spec 0212's u32
NodeSpan field never receives an out-of-range tag. Two of the three
narrowing sites merely hid that behind a u64 and are now u32 end to
end; extract's message_payload_range, which re-parses a raw tag with no
span to lean on, gained a real check.

Two shared test fixtures now keep their temp descriptor set, since
hashing it is a fresh read from disk.

docs/protolens/audit-quality.md

+++ protolens/src/export_descriptor.rs
@@ -21,7 +21,10 @@ use std::collections::HashSet;
 /// `referenced_file` is `Some` only for `Message`/`Enum` fields, and
 /// feeds `build`'s declared-dependency list (G6d).
 pub struct ResolvedField {
-    pub number: u64,
+    /// A `u32`, matching `NodeSpan::field_number`, which is where every
+    /// one of these comes from — so `build`'s narrowing to
+    /// `FieldDescriptorProto`'s `i32` is exact (spec 0212).
+    pub number: u32,
     pub name: String,
     pub label: Label,
     pub r#type: Type,
